feat(extension): show recorded event count on event drawer toggle
Share recorded events state between the toolbox and event drawer via
useRecordedEvents so the toolbar badge stays in sync with the list.

Closes grafana/k6-studio#657
