+++ extension/src/frontend/view/EventDrawer.tsx
@@ -1,7 +1,6 @@
 import { css } from '@emotion/react'
 import * as Dialog from '@radix-ui/react-dialog'
 import { XIcon } from 'lucide-react'
-import { useEffect, useState } from 'react'
 
 import { BrowserEventList } from '@/components/BrowserEventList'
 import { useContainerElement } from '@/components/primitives/ContainerProvider'
@@ -11,61 +10,21 @@ import { RecordingContext } from '@/views/Recorder/RecordingContext'
 
 import { useStudioClient } from './StudioClientProvider'
 
-function useRecordedEvents() {
-  const client = useStudioClient()
-
-  const [events, setEvents] = useState<BrowserEvent[]>([])
-
-  useEffect(() => {
-    return client.on('events-recorded', (event) => {
-      setEvents((prev) => [...prev, ...event.data.events])
-    })
-  }, [client])
-
-  useEffect(() => {
-    return client.on('events-loaded', (event) => {
-      setEvents(event.data.events)
-    })
-  }, [client])
-
-  useEffect(() => {
-    client.send({
-      type: 'load-events',
-    })
-  }, [client])
-
-  useEffect(() => {
-    // We reload the list of events whenever the page is shown from the
-    // back/forward cache to make sure we have the latest state.
-    function handlePageShow(event: PageTransitionEvent) {
-      if (event.persisted) {
-        client.send({
-          type: 'load-events',
-        })
-      }
-    }
-
-    window.addEventListener('pageshow', handlePageShow)
-
-    return () => {
-      window.removeEventListener('pageshow', handlePageShow)
-    }
-  }, [client])
-
-  return events
-}
-
 interface EventDrawerProps {
   open: boolean
   editing: boolean
+  events: BrowserEvent[]
   onOpenChange: (open: boolean) => void
 }
 
-export function EventDrawer({ open, editing, onOpenChange }: EventDrawerProps) {
+export function EventDrawer({
+  open,
+  editing,
+  events,
+  onOpenChange,
+}: EventDrawerProps) {
   const client = useStudioClient()
   const container = useContainerElement()
-
-  const events = useRecordedEvents()
 
   const handleHighlight = (selector: NodeSelector | null) => {
     client.send({
