+++ extension/src/frontend/view/hooks/useRecordedEvents.ts
@@ -1,0 +1,47 @@
+import { useEffect, useState } from 'react'
+
+import { BrowserEvent } from '@/schemas/recording'
+
+import { useStudioClient } from '../StudioClientProvider'
+
+export function useRecordedEvents() {
+  const client = useStudioClient()
+
+  const [events, setEvents] = useState<BrowserEvent[]>([])
+
+  useEffect(() => {
+    return client.on('events-recorded', (event) => {
+      setEvents((prev) => [...prev, ...event.data.events])
+    })
+  }, [client])
+
+  useEffect(() => {
+    return client.on('events-loaded', (event) => {
+      setEvents(event.data.events)
+    })
+  }, [client])
+
+  useEffect(() => {
+    client.send({
+      type: 'load-events',
+    })
+  }, [client])
+
+  useEffect(() => {
+    function handlePageShow(event: PageTransitionEvent) {
+      if (event.persisted) {
+        client.send({
+          type: 'load-events',
+        })
+      }
+    }
+
+    window.addEventListener('pageshow', handlePageShow)
+
+    return () => {
+      window.removeEventListener('pageshow', handlePageShow)
+    }
+  }, [client])
+
+  return events
+}
